requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.5.0; python_version >= '3.7'
attrs==21.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
cfgv==3.3.1; python_full_version >= '3.6.1'
click==8.0.3; python_version >= '3.6'
distlib==0.3.4
django-filter==21.1
django==3.2.12
djangorestframework==3.13.1
factory-boy==3.2.1; python_version >= '3.6'
faker==12.3.0; python_version >= '3.6'
filelock==3.4.2; python_version >= '3.7'
flake8==4.0.1
identify==2.4.9; python_version >= '3.7'
inflection==0.5.1; python_version >= '3.5'
iniconfig==1.1.1
mccabe==0.6.1
nodeenv==1.6.0
packaging==21.3; python_version >= '3.6'
pipfile-sort==0.2.2
platformdirs==2.5.0; python_version >= '3.7'
plette==0.2.3; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
pluggy==1.0.0; python_version >= '3.6'
pre-commit==2.17.0
psycopg2==2.9.3
py==1.11.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pycodestyle==2.8.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pyflakes==2.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyparsing==3.0.7; python_version >= '3.6'
pytest-django==4.5.2
pytest-factoryboy==2.1.0
pytest==7.0.1
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2021.3
pyyaml==6.0; python_version >= '3.6'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.2; python_version >= '3.5'
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
tomli==2.0.1; python_version >= '3.7'
tomlkit==0.9.2; python_version >= '3.6' and python_version < '4.0'
virtualenv==20.13.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
